tests/end2end: migrate tests to safe server code: UC06
Relates to #925
