--- tests/end2end/document/UC06_create_requirement/06_03_creating_section_after_requirement/test_06_03_creating_section_after_requirement.py
@@ -1,98 +1,0 @@
-import filecmp
-import os
-import shutil
-
-from selenium.webdriver.common.by import By
-from seleniumbase import BaseCase
-
-from tests.end2end.server import SDocTestServer
-
-path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
-
-
-class Test06CreateSectionAfterRequirement(BaseCase):
-    def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
-
-        test_server = SDocTestServer.create(path_to_sandbox)
-        shutil.copyfile(
-            os.path.join(path_to_this_test_file_folder, "document.sdoc"),
-            os.path.join(path_to_sandbox, "document.sdoc"),
-        )
-
-        test_server.run()
-
-        self.open(test_server.get_host_and_port())
-
-        self.assert_text("Document 1")
-        self.assert_text("PROJECT INDEX")
-
-        self.click_xpath('//*[@data-testid="tree-file-link"]')
-
-        self.assert_text("Hello world!")
-
-        # Requirement
-
-        self.hover_and_click(
-            hover_selector="(//sdoc-node)[1]",
-            click_selector=(
-                '(//sdoc-node)[1]//*[@data-testid="node-menu-handler"]'
-            ),
-            hover_by=By.XPATH,
-            click_by=By.XPATH,
-        )
-        self.click(
-            selector=(
-                "(//sdoc-node)[1]"
-                '//*[@data-testid="node-add-requirement-first-action"]'
-            ),
-            by=By.XPATH,
-        )
-
-        self.type(
-            "//*[@id='requirement[TITLE]']", "Requirement title #1", by=By.XPATH
-        )
-        self.type(
-            "//*[@id='requirement[STATEMENT]']",
-            "Requirement statement #1.",
-            by=By.XPATH,
-        )
-
-        self.click_xpath('//*[@data-testid="form-submit-action"]')
-
-        self.assert_text("1. Requirement title #1")
-
-        # Section
-
-        self.hover_and_click(
-            hover_selector="(//sdoc-node)[2]",
-            click_selector=(
-                '(//sdoc-node)[2]//*[@data-testid="node-menu-handler"]'
-            ),
-            hover_by=By.XPATH,
-            click_by=By.XPATH,
-        )
-        self.click(
-            selector=(
-                "(//sdoc-node)[2]"
-                '//*[@data-testid="node-add-section-below-action"]'
-            ),
-            by=By.XPATH,
-        )
-
-        self.type("#section_title", "Section title")
-        self.type("#section_content", "Section statement.")
-
-        self.click_xpath('//*[@data-testid="form-submit-action"]')
-
-        self.assert_text("2. Section title")
-
-        assert os.path.exists(os.path.join(path_to_sandbox, "document.sdoc"))
-        assert filecmp.cmp(
-            os.path.join(path_to_sandbox, "document.sdoc"),
-            os.path.join(
-                path_to_this_test_file_folder, "document.expected.sdoc"
-            ),
-        )
