+++ tests/end2end/document/UC06_create_requirement/UC06_G1_validations/UC06_G1_T01_requirement_with_no_statement/expected_output/document.sdoc
