+++ tests/end2end/document/UC06_create_requirement/UC06_G1_validations/UC06_G1_T02_malformed_rst_statement/expected_output/document.sdoc
@@ -1,0 +1,6 @@
+[DOCUMENT]
+TITLE: Document 1
+
+[FREETEXT]
+Hello world!
+[/FREETEXT]
