+++ tests/end2end/document/UC06_create_requirement/UC06_T01_cancel_new_requirement/expected_output/document.sdoc
