+++ tests/end2end/document/UC06_create_requirement/UC06_T01_cancel_new_requirement/input/document.sdoc
