+++ tests/end2end/document/UC06_create_requirement/UC06_T08_sanitize_trailing_symbols/input/document.sdoc
